refactor: organize blog post utils
